+++ tests/testthat/test-h3_get_resolution.R
@@ -9,7 +9,7 @@ testthat::skip_if_not_installed("duckdb")
 testthat::skip_if_not_installed("duckspatial")
 
 ## create duckdb connection
-conn_test <- duckh3::ddbh3_create_conn()
+conn_test <- duckh3::ddbh3_default_conn()
 
 ## Load example data
 points_tbl <- read.csv(
@@ -192,8 +192,9 @@ describe("errors", {
   })
   
   it("requires connection when using table names", {
-    expect_warning(
-      expect_true(is.na(ddbh3_get_resolution("test_data_tbl", conn = NULL)))
+    expect_error(
+      ddbh3_get_resolution("test_data_tbl", conn = NULL),
+      "If `x` is a table, `conn` cannot be NULL"
     )
   })