+++ tests/testthat/test-h3_vertex.R
@@ -9,7 +9,7 @@ testthat::skip_if_not_installed("duckdb")
 testthat::skip_if_not_installed("duckspatial")
 
 ## create duckdb connection
-conn_test <- duckh3::ddbh3_create_conn()
+conn_test <- duckh3::ddbh3_default_conn()
 
 ## Load example data
 test_data <- read.csv(
@@ -234,8 +234,8 @@ describe("errors", {
   })
 
   it("requires connection when using table names", {
-    expect_warning(
-      expect_true(is.na(ddbh3_h3_to_vertex("test_data_tbl", conn = NULL)))
+    expect_error(
+      ddbh3_h3_to_vertex("test_data_tbl", conn = NULL)
     )
   })
 
@@ -450,8 +450,8 @@ describe("errors", {
   })
 
   it("requires connection when using table names", {
-    expect_warning(
-      expect_true(is.na(ddbh3_vertex_to_lon("test_data_tbl", conn = NULL)))
+    expect_error(
+      ddbh3_vertex_to_lon("test_data_tbl", conn = NULL)
     )
   })
 
@@ -666,8 +666,8 @@ describe("errors", {
   })
 
   it("requires connection when using table names", {
-    expect_warning(
-      expect_true(is.na(ddbh3_vertex_to_lat("test_data_tbl", conn = NULL)))
+    expect_error(
+      ddbh3_vertex_to_lat("test_data_tbl", conn = NULL)
     )
   })